requirements.txt
numpy==2.3.1
pandas==2.3.0
pygame==2.6.1
scikit_learn==1.7.0
matplotlib-3.10.3
seaborn-0.13.2
yfinance==0.2.64
